requirements.txt
streamlit>=1.44.0
scikit-learn>=1.5.0
pypdf>=5.4.0
python-docx>=1.1.2
